+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 1.3.23
+
+added gnss3 orbit option for testing. eventually will use this for gnss.
+it downloads multi-GNSS sp3files directly from GFZ instead of from CDDIS.
+
+
 ## 1.3.22
 allows gzip RINEX 2.11 files in local directory